chore: have plan to publish stemlib as crate, first need to complete the rmq prodcons streaming and rpc
