+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ovm"
+name = "stem"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,12 +7,12 @@ edition = "2021"
 
 [workspace]
 members = [ 
-  "ovmlib" # ovmlib has separate Cargo.toml
+  "stemlib" # stemlib has separate Cargo.toml
 ]
 
 [dependencies]
 crypter = { path = "crypter", name = "crypter" }
-ovmlib = { path = "ovmlib", name = "ovmlib"}
+stemlib = { path = "stemlib", name = "stemlib"}
 salvo = { version = "0.73.0", features = ["full"] }
 actix = "0.13.5"
 actix-web = "4"
@@ -57,7 +57,7 @@ wasmer = "4.4.0"
 wasmer-wasi = "3.1.1"
 
 [[bin]]
-name = "ovm"
+name = "stem"
 path = "src/main.rs"
 
 [[bin]]
