+++ stemlib/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ovmlib"
+name = "stemlib"
 version = "0.1.0"
 edition = "2021"
 
@@ -42,7 +42,7 @@ indexmap = "2.4"
 # we need lib.rs and crate type of cdylib for none binary like .wasm, .so, .dylib and .dll files or for tests
 crate-type = ["cdylib", "rlib"]
 path = "src/lib.rs"
-name = "ovmlib"
+name = "stemlib"
 
 [profile.release]
 panic = "abort"
