+++ stemlib/src/task.rs
